git_pylint/requirements.txt
astroid==2.2.5
backports.functools-lru-cache==1.4
configparser==3.5.0
enum34==1.1.6
gitdb2==2.0.2
GitPython==2.1.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
protobuf==3.3.0
pylint==2.3.1
singledispatch==3.4.0.3
six==1.11.0
smmap2==2.0.3
wrapt==1.10.10
typed_ast==1.4.0